requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
cloudinary==1.40.0
django-cloudinary-storage==0.3.0
Pillow==10.3.0
django-filter==24.2
